Change email login to national ID login in authentication system
The diff includes changes to the `LoginRequest` class in the `app/Http/Requests/Auth` directory and the `Login.vue` component in the `resources/js/Pages/Auth` directory.

In `LoginRequest`, the `email` field has been changed to `national_id` in the `rules` method, and the `attempt` and `throttleKey` methods have been updated to use `national_id` instead of `email`.

In `Login.vue`, the `email` field has been changed to `national_id` in the `form` object and the `InputLabel` and `InputError` components. The `type` attribute of the `TextInput` component has also been changed from `email` to `text`.